Renamed power-indicator.c to main.c (#18)
In hindsight main.c is the way to go for the entry point sourcefile. It
is what most people will look for (I suspect) when the look at the
source code for the first time. Power-indicator.c does not jump out as
much coming back to the code base after a while.