Updated indicator to use an LED matrix instead of LED strip (#19)
In order to display more information in an easy to understand way we are
switching to a matrix LED array. This will allow us to have multiple
rows which we can use to communicate different information more clearly.
Additionally the first row is reserved as a status bar which can be
broken up into different segments to display things like Wi-Fi
connection status.

Additionally, the network module was updated to return false if the
connection failed.